added security config, added email field,changed id from long to int

+++ GRSpring/src/main/java/com/revature/service/UserService.java
@@ -5,9 +5,9 @@ import com.revature.models.User;
 public interface UserService {
 	
 	User save(User user);
-	User findById(long id);
+	User findById(int id);
 	User findByUsername(String username);
-	void deleteUser(long id);
+	void deleteUser(int id);
 
 
 }
